requirements.txt
Flask==0.12.2
Flask-HTTPAuth==3.2.3
python-dotenv==0.10.1
